refactor: extract customer-workset, dashboard, billing routes from workset_api.py
- Created daylib/routes/customer_worksets.py (workset CRUD, logs, metrics, snakemake)
- Created daylib/routes/dashboard.py (activity, cost history, cost breakdown)
- Created daylib/routes/billing.py (billing summary, workset billing)
- Updated daylib/routes/__init__.py with new router exports
- Added include_router() calls in workset_api.py
- Fixed test patch target for boto3 in moved delete route
- workset_api.py: 3972 → 2287 lines (~1685 line reduction)
- All 769 tests pass

Agent-Id: agent-3874f94b-f39d-4ee1-9e59-7746c72f1de4
